requirements.txt
discord.py[voice]>=2.3.0
yt-dlp>=2023.7.6
python-dotenv>=1.0.0
colorlog>=6.7.0
PyNaCl>=1.5.0